Set and test Message options
To differentiate between read and write operations, Gitaly and its
clients will need annotations. This change includes a way to do so, on
a message level.

Another option, would be to annotate the method. This might have
advantages, and instinctively it seems better to me.

For sake of the repository size, required packages for testing are
assumed to be present in the $GOPATH.